TIP-459: Merge ArrayReaders
