+++ src/Pim/Bundle/ConnectorBundle/Resources/config/readers.yml
@@ -5,13 +5,11 @@ parameters:
     pim_connector.reader.database.product.class: Pim\Component\Connector\Reader\Database\ProductReader
 
     pim_connector.reader.file.csv.class:               Pim\Component\Connector\Reader\File\Csv\Reader
-    pim_connector.reader.file.csv_array.class:         Pim\Component\Connector\Reader\File\Csv\ArrayReader
     pim_connector.reader.file.csv_variant_group.class: Pim\Component\Connector\Reader\File\Csv\VariantGroupReader
     pim_connector.reader.file.csv_association.class:   Pim\Component\Connector\Reader\File\Csv\ProductAssociationReader
     pim_connector.reader.file.csv_product.class:       Pim\Component\Connector\Reader\File\Csv\ProductReader
 
     pim_connector.reader.file.xlsx.class:         Pim\Component\Connector\Reader\File\Xlsx\Reader
-    pim_connector.reader.file.xlsx_array.class:   Pim\Component\Connector\Reader\File\Xlsx\ArrayReader
     pim_connector.reader.file.xlsx_product.class: Pim\Component\Connector\Reader\File\Xlsx\ProductReader
 
     pim_connector.reader.file.yaml.class: Pim\Component\Connector\Reader\File\Yaml\Reader
@@ -19,6 +17,7 @@ parameters:
     pim_connector.reader.file.file_iterator.class:          Pim\Component\Connector\Reader\File\FileIterator
     pim_connector.reader.file.file_iterator_factory.class:  Pim\Component\Connector\Reader\File\FileIteratorFactory
     pim_connector.reader.file.media_path_transformer.class: Pim\Component\Connector\Reader\File\MediaPathTransformer
+    pim_connector.reader.file.array_reader.class:           Pim\Component\Connector\Reader\File\ArrayReader
 
 services:
     pim_connector.reader.database.product:
